requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-genai>=1.10.0
qdrant-client>=1.14.0
python-multipart>=0.0.9
